Back/requirements.txt
fastapi==0.115.11
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.39
uvicorn==0.34.0
aiomysql==0.2.0
pipreqs
pip-check-reqs
passlib[bcrypt]
bcrypt==4.0.1
python-jose==3.3.0
cryptography
pydantic[email]==2.10.6
apscheduler==3.10.4
asyncpg==0.28.0
uvicorn
gunicorn
openai>=1.0.0